requirements.txt
crewai==0.80.0
langchain==0.3.0
langchain-community==0.3.0
mistralai==1.2.0
langchain-mistralai==0.2.0
qdrant-client==1.9.0
sentence-transformers==2.7.0
transformers==4.40.2
torch==2.2.2
firecrawl-py==1.4.0
python-dotenv==1.0.0
pydantic==2.7.0
pyyaml==6.0.1
requests==2.31.0
reportlab==4.2.0
schedule==1.2.1
rich==13.7.0
tenacity==8.3.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
urllib3==2.0.7
charset-normalizer==3.3.2
